requirements.txt
google-generativeai
dataclasses_json
python-dotenv
bitsandbytes
sentence-transformers
transformers
accelerate
numpy
fire